Add option for enabling/disabling automatic GCM duplicate deletion
Use deleteDuplicatesAutomatically option under exports['gcm'] to control
whether duplicates should be deleted.

+++ lib/pushservices/gcm.coffee
@@ -8,6 +8,7 @@ class PushServiceGCM
 
     constructor: (conf, @logger, tokenResolver) ->
         conf.concurrency ?= 10
+        conf.deleteDuplicatesAutomatically ?= false
         @driver = new gcm.Sender(conf.key)
         @multicastQueue = {}
 
@@ -54,11 +55,12 @@ class PushServiceGCM
     handleResult: (result, subscriber) ->
         if result.messageId or result.message_id
             if result.registration_id
-                # GCM sends registration_id if there is duplicate subscription with same device/app ID.
-                # Remove subscriber to avoid duplicate notifications. GCM does not send this to latest
-                # subscription so there is no race condition where both subscriptions will be deleted.
-                @logger?.warn("GCM Removing subscriber #{subscriber.id} of canonical #{result.canonical_id}")
-                subscriber.delete()
+                if @conf.deleteDuplicatesAutomatically
+                    # GCM sends registration_id if there is duplicate subscription with same device/app ID.
+                    # Remove subscriber to avoid duplicate notifications. GCM does not send this to latest
+                    # subscription so there is no race condition where both subscriptions will be deleted.
+                    @logger?.warn("GCM Removing subscriber #{subscriber.id} of canonical #{result.canonical_id}")
+                    subscriber.delete()
         else
             error = result.error or result.errorCode
             if error is "NotRegistered" or error is "InvalidRegistration"
